+++ src/main/java/com/hyun/oauthboard/domain/board/dto/BoardIdReponse.java
@@ -1,0 +1,11 @@
+package com.hyun.oauthboard.domain.board.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class BoardIdReponse {
+
+    private Long boardId;
+}
