+++ src/main/java/com/hyun/oauthboard/domain/board/dto/BoardResponse.java
@@ -1,0 +1,26 @@
+package com.hyun.oauthboard.domain.board.dto;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import java.time.LocalDateTime;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+@AllArgsConstructor
+public class BoardResponse {
+
+    private Long boardId;
+    private Long memberId;
+    private String boardTitle;
+    private String boardContent;
+    private Long views;
+    private Long likes;
+
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm")
+    private LocalDateTime createdAt;
+
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm")
+    private LocalDateTime updatedAt;
+}
