+++ src/main/java/com/hyun/oauthboard/domain/board/entity/BoardView.java
@@ -1,0 +1,48 @@
+package com.hyun.oauthboard.domain.board.entity;
+
+import com.hyun.oauthboard.domain.member.entity.Member;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import java.time.LocalDateTime;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+@Table(name = "board_view")
+@Entity
+@Builder
+@EntityListeners(AuditingEntityListener.class)
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class BoardView {
+
+    @Id
+    @Column(name = "view_id")
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long viewId;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "viewer_id", nullable = false)
+    private Member viewer;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "board_id", nullable = false)
+    private Board board;
+
+    @CreatedDate
+    @Column(name = "created_at", updatable = false)
+    private LocalDateTime createdAt;
+}
