+++ src/main/java/com/hyun/oauthboard/domain/board/entity/model/BoardEntityModel.java
@@ -1,0 +1,33 @@
+package com.hyun.oauthboard.domain.board.entity.model;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.MappedSuperclass;
+import jakarta.persistence.PrePersist;
+import jakarta.persistence.PreUpdate;
+import java.time.LocalDateTime;
+import lombok.Getter;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+@MappedSuperclass
+@Getter
+@EntityListeners(AuditingEntityListener.class)
+public abstract class BoardEntityModel {
+
+    @Column(name = "created_at")
+    private LocalDateTime createdAt;
+
+    @Column(name = "updated_at")
+    private LocalDateTime updatedAt;
+
+    @PrePersist
+    void onPrePersist() {
+        this.createdAt = LocalDateTime.now();
+        this.updatedAt = createdAt;
+    }
+
+    @PreUpdate
+    void onPreUpdate() {
+        this.updatedAt = LocalDateTime.now();
+    }
+}
