+++ src/main/java/com/hyun/oauthboard/domain/board/repository/BoardCustomRepository.java
@@ -1,0 +1,8 @@
+package com.hyun.oauthboard.domain.board.repository;
+
+import com.hyun.oauthboard.domain.board.dto.BoardResponse;
+
+public interface BoardCustomRepository {
+
+    BoardResponse findBoardLeftJoinViewsAndLikes(Long boardId);
+}
