+++ src/main/java/com/hyun/oauthboard/domain/board/repository/BoardRepository.java
@@ -1,0 +1,10 @@
+package com.hyun.oauthboard.domain.board.repository;
+
+import com.hyun.oauthboard.domain.board.entity.Board;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface BoardRepository extends JpaRepository<Board, Long>, BoardCustomRepository {
+
+//    @Query("SELECT DISTINCT b FROM Board b LEFT JOIN FETCH b.likes WHERE b.boardId = :boardId")
+//    Optional<Board> findById(@Param("boardId") Long boardId);
+}
