--- src/main/java/com/hyun/oauthboard/domain/dto/board/BoardIdReponse.java
@@ -1,11 +1,0 @@
-package com.hyun.oauthboard.domain.dto.board;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class BoardIdReponse {
-
-    private Long boardId;
-}
