--- src/main/java/com/hyun/oauthboard/domain/dto/github/GithubMemberInfo.java
@@ -1,17 +1,0 @@
-package com.hyun.oauthboard.domain.dto.github;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class GithubMemberInfo {
-    private String login;
-    private Long id;
-
-    @JsonProperty("avatar_url")
-    private String avatarUrl;
-
-    private String name;
-}
