--- src/main/java/com/hyun/oauthboard/domain/dto/github/OAuthAccessTokenRequest.java
@@ -1,24 +1,0 @@
-package com.hyun.oauthboard.domain.dto.github;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class OAuthAccessTokenRequest {
-
-    @JsonProperty("client_id")
-    private String clientId;
-
-    @JsonProperty("client_secret")
-    private String clientSecret;
-    private String code;
-
-    public OAuthAccessTokenRequest(final String clientId, final String clientSecret, final String code) {
-        this.clientId = clientId;
-        this.clientSecret = clientSecret;
-        this.code = code;
-    }
-
-}
