--- src/main/java/com/hyun/oauthboard/domain/dto/github/OAuthTokensResponse.java
@@ -1,14 +1,0 @@
-package com.hyun.oauthboard.domain.dto.github;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class OAuthTokensResponse {
-
-    @JsonProperty("access_token")
-    private String accessToken;
-
-}
