--- src/main/java/com/hyun/oauthboard/domain/dto/jwt/JwtToken.java
@@ -1,14 +1,0 @@
-package com.hyun.oauthboard.domain.dto.jwt;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-
-@Builder
-@Data
-@AllArgsConstructor
-public class JwtToken {
-
-    private String accessToken;
-    private String refreshToken;
-}
