+++ src/main/java/com/hyun/oauthboard/domain/member/repository/MemberRepository.java
@@ -1,0 +1,10 @@
+package com.hyun.oauthboard.domain.member.repository;
+
+import com.hyun.oauthboard.domain.member.entity.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface MemberRepository extends JpaRepository<Member, Long> {
+
+}
