+++ src/main/java/com/hyun/oauthboard/jwt/JwtBlacklistService.java
@@ -1,0 +1,8 @@
+package com.hyun.oauthboard.jwt;
+
+public interface JwtBlacklistService {
+
+    void addBlackList(String token);
+
+    boolean isBlackList(String token);
+}
