+++ src/main/java/com/hyun/oauthboard/jwt/JwtToken.java
@@ -1,0 +1,14 @@
+package com.hyun.oauthboard.jwt;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+
+@Builder
+@Data
+@AllArgsConstructor
+public class JwtToken {
+
+    private String accessToken;
+    private String refreshToken;
+}
