--- src/main/java/com/hyun/oauthboard/repository/BoardCustomRepository.java
@@ -1,8 +1,0 @@
-package com.hyun.oauthboard.repository;
-
-import com.hyun.oauthboard.domain.dto.board.BoardResponse;
-
-public interface BoardCustomRepository {
-
-    BoardResponse findBoardLeftJoinViewsAndLikes(Long boardId);
-}
