--- src/main/java/com/hyun/oauthboard/repository/MemberRepository.java
@@ -1,10 +1,0 @@
-package com.hyun.oauthboard.repository;
-
-import com.hyun.oauthboard.domain.entity.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface MemberRepository extends JpaRepository<Member, Long> {
-
-}
