--- src/main/java/com/hyun/oauthboard/service/JwtBlacklistService.java
@@ -1,8 +1,0 @@
-package com.hyun.oauthboard.service;
-
-public interface JwtBlacklistService {
-
-    void addBlackList(String token);
-
-    boolean isBlackList(String token);
-}
